Fix setup script, worlds and plugin bug
Fix errors in the setup script.

cuberite stores multiple values with the same key in the `settings.ini`.
pterodactyl's ini parser does not allow this and takes only the last (?)
and removes the other entries.

The easy fix is to not edit the `settings.ini` and have the correct
values in the template.
This could run into problems if same servers on different node use a
different port.